feat: add the 'New Language Server' menu on explorer right-click
