+++ src/main/resources/messages/LanguageServerBundle.properties
@@ -100,7 +100,7 @@ action.lsp.console.explorer.export.servers.notification.message.error=Language s
 action.lsp.console.explorer.export.servers.notification.message.single=Successfully exported 1 language server.
 action.lsp.console.explorer.export.servers.notification.message.multi=Successfully exported {0} language servers.
 action.lsp.console.folding.text=Collapse/Expand All
-action.lsp.console.new.language.server.text=New language server
+action.lsp.console.new.language.server.text=New Language Server
 action.lsp.console.new.language.server.description=Open dialog to create a new language server
 action.lsp.detail.apply.text=Apply
 action.lsp.detail.apply.description=Apply changes for this language server settings
